+++ src/entmoot/routes/__init__.py
@@ -1,0 +1,4 @@
+from .attributes import AttributeController
+from .entities import EntityController
+
+__all__ = ["EntityController", "AttributeController"]